Enhance NOAA Daily Data Ingestion and GTFS-RT Snapshot Workflows
- Updated `noaa_daily.py` to use nullable integer type for snow_day column to handle NULLs while preserving integer format.
- Added GitHub Actions workflows for hourly loading of GTFS-RT snapshots into BigQuery and for capturing hourly GTFS-RT snapshots.
- Documented the GTFS-RT snapshot strategy, detailing data collection architecture, workflow architecture, cost analysis, and data quality benefits.
- Created a comprehensive NOAA weather data strategy document outlining data architecture, ingestion strategy, cost efficiency, data quality expectations, and troubleshooting steps.
- Implemented a backfill script for NOAA weather data from January 2024 to September 2025, ensuring historical data completeness.

+++ src/whylinedenver/ingest/noaa_daily.py
@@ -261,7 +261,10 @@ def build_dataframe(
     df = pd.DataFrame(rows)
     df = _normalize_units(df)
     df["tavg_c"] = df["tavg_c"].combine_first((df["tmin_c"] + df["tmax_c"]) / 2)
-    df["snow_day"] = df["snow_mm"].apply(lambda x: None if pd.isna(x) else int(x >= 1.0))
+    # Use Int64 (nullable integer) to preserve integer format even with NULLs
+    df["snow_day"] = (
+        df["snow_mm"].apply(lambda x: None if pd.isna(x) else int(x >= 1.0)).astype("Int64")
+    )
     df["precip_bin"] = df["precip_mm"].apply(_precip_category)
     for column in ["snow_mm", "precip_mm", "tmin_c", "tmax_c", "tavg_c"]:
         if column in df:
